Make the `dtype` node parameter reach the node state and run output
`Reservoir` / `LocalPlasticityReservoir` build their weights with the
configured `dtype`, but `initialize()` seeded `state` with a plain
`np.zeros(...)` (always float64). The float64 state then upcast every
`_step` (`W @ s`), so a node built with `dtype=np.float32` stored float32
weights yet computed and returned float64.

- `Reservoir.initialize` / `LocalPlasticityReservoir.initialize`: seed the
  state with `dtype=self.dtype`.
- `Node._run`: allocate the output buffer at `state["out"].dtype` so
  `run()` returns the node's dtype instead of always float64.
- `Node.reset`: preserve each state entry's dtype instead of forcing
  float64.

`ES2N`, `IPReservoir` and `LIF` carry further float64 internals
(`mat_gen.orthogonal` does not honour its `dtype` arg; `IPReservoir.a/.b`;
`LIF._step`'s `np.where(cond, 1.0, 0.0)` spike cast) and their state stays
float64 after this change -- tracked separately.

New tests in `reservoirpy/tests/test_dtype.py`; default (float64) behaviour
is unchanged.

+++ reservoirpy/nodes/local_plasticity_reservoir.py
@@ -322,8 +322,8 @@ class LocalPlasticityReservoir(TrainableNode):
             )
 
         self.state = {
-            "out": np.zeros((self.units,)),
-            "internal": np.zeros((self.units,)),
+            "out": np.zeros((self.units,), dtype=self.dtype),
+            "internal": np.zeros((self.units,), dtype=self.dtype),
         }
 
         self.initialized = True
